Eval() is the only way it works
If you can figure out why that is I'd be most delighted. Cannot store document.getElementById() and call it within function() without using string method "". The program will not toggle the button. ParseInt might work or numbers? But will it replace a string and run the function and then convert to a value? Is that value machine code? Or are we stuck with a useless number that won't affect the style of the document? Hmmmm